<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
主要修改： 修改了grid size最大值的设置，原本develop中的grid size最大值为8，并行block数目太少.

benchmark 测试数据：
[benchmark script](https://github.com/PaddlePaddle/benchmark/blob/master/api/tests_v2/yolo_box.py)
This PR
| 输入(x_shape) | Kernel average | 
|---|---|
| [32, 21, 13, 13] | 9.9110us | 
| [128, 21, 26, 26] | 56.239us | 

develop:

| 输入(x_shape) | Kernel average | 
|---|---|
| [32, 21, 13, 13] | 20.555us | 
| [128, 21, 26, 26] | 333.00us |